Use bash built-in sleep when available
The existing "watch tomate" script uses the `sleep` executable to wait
between two invocations of `pgrep`, to avoid calling the process
needlessly quickly. This leads to a new process being created and
killed every few seconds.

Instead, when available, use the bash built-in sleep, which achieves
the same thing without spawning a new sub-process on every iteration
of the loop, effectively halving the number of processes we spawn.

Note: the loadable sleep path is currently hard-coded to work on Arch
Linux, but we do not search other possible paths, so your system might
have the loadable sleep available elsewhere but the script still won't
use it.

Signed-off-by: Antoine Busque <antoinebusque@gmail.com>